fix: check stuckPlayer before handleNoLegalPlay to show correct message
Previously, we checked state.stuckPlayer after calling handleNoLegalPlay(),
which sets stuckPlayer if it's the first time. This caused the check to
always see stuckPlayer as set, so we always showed 'No available moves'
instead of 'No available moves, last round!' for the first stuck player.

Now we check stuckPlayer before calling handleNoLegalPlay() to correctly
determine if this is the first player to get stuck or a subsequent player.